requirements.txt
asgiref==3.10.0
contourpy==1.3.3
crispy-bootstrap5==2025.6
cycler==0.12.1
diff-match-patch==20241021
Django==5.2.8
django-crispy-forms==2.5
django-import-export==4.3.13
django-widget-tweaks==1.5.0
fonttools==4.60.1
kiwisolver==1.4.9
matplotlib==3.10.7
narwhals==2.11.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.4.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
tablib==3.9.0
tzdata==2025.2
